Introduce TaskTemplate::finishTaskQueued
We invented tasks as a replacement for state machines and they work fine in
several places.

Now we learned that we missed important: State machines change state queued but
tasks - often part of a sequential sequence - are direct connected and transit
immediately.
So when replacing state machines we can run into trouble: E.g in a sequential task
sequence the next task can start before events generated by previous task
are handled.
Up to now this has not been a problem but now we try to replace a state machine
in zera-classes' BaseModule and do not get tests to pass.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>